Add 'no platform sounds' argument to NFCReaderMode
