+++ lib/nfc_in_flutter.dart
@@ -1,4 +1,3 @@
 export './src/api.dart';
 export "./src/messages.dart";
 export "./src/exceptions.dart";
-export "./src/reader_modes.dart";
